Replace setPhases with setPhaseShifts, xShift, yShift
This removes the ability to explicitly set the computed phase values
for the current frame. Instead, new `xShift` and `yShift` properties
are added, along with convenience functions to set them, which act
as a direct phase shift for the resulting pattern. This is more
flexible, as it doesn't require adjusting internal state and allows
the phase offsets to be additive with the underlying animation values.