Update pattern API with support for more pattern types
This is a major overhaul of the core API for declaring patterns.
Four types are now supported:

1. Bit Patterns: An array of 8 numbers describing the bitmap for each
   row, with optional 8 for alpha. (The standard used by setPattern()
   in the SDK and previously supported.)
2. Dither Patterns: A table containing `ditherType` and optional
   `alpha` and `color` values to generate a pattern with. Formerly,
   those properties were set independently on the EasyPattern; now
   they are passed as one to the `pattern` property.
3. Images: An 8x8 pixel `playdate.graphics.image`
4. Image Tables: An 8x8 pixel `playdate.graphics.imagetable` with an
   arbitrary number of images which get shown in sequence for
   `tickDuration` seconds.

These types are supported for both the `pattern` and `bgPattern`
properties. Accordingly, the `setPattern()` and `setBackgroundPattern()`
functions are overloaded to accept any of these types.

The `alpha` and `ditherType` properties on the `EasyPattern` itself
still exist, however they are now repurposed to define the transparency
effect for the overall pattern, in the spirit of `drawFaded()`.

Includes a fully updated test suite, as well as revised README with
additional examples to highlight the new capabilities.